% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{Intervals}
\alias{Intervals}
\title{Example genomic intervals}
\usage{
data(Intervals)
}
\value{
A GRanges object with two rows
}
\description{
This dataset contains peaks from an in-house EBF1 ChIP-seq
}
\details{
\itemize{
\item Intervals GRanges object containing EBF1 peaks
}
}
\keyword{datasets}

