% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{makebedtable}
\alias{makebedtable}
\title{Make HTML pages for interval files or GRanges.}
\usage{
makebedtable(grangesObject, name, basedirectory)
}
\arguments{
\item{grangesObject}{A GRanges object.}

\item{name}{Character of the name for Interval HTML report.}

\item{basedirectory}{Character of the directory to which HTML report is writen.}
}
\description{
Creates HTML pages for interval files or GRanges (Tracktables Interval Report).
}
\examples{
fileLocations <- system.file("extdata",package="tracktables")

bigwigs <- dir(fileLocations,pattern="*.bw",full.names=TRUE)

intervals <- dir(fileLocations,pattern="*.bed",full.names=TRUE)
intervalGRanges <- tracktables:::GetGRanges(intervals[1])
htmlpage <- makebedtable(intervalGRanges,"EBF_PeaksTable.html",getwd())
}
\author{
Thomas Carroll
}

