% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{plotRegion}
\alias{plotRegion}
\alias{plotRegion,ChIPprofile-method}
\alias{plotRegion.ChIPprofile}
\title{Plot regions}
\usage{
\S4method{plotRegion}{ChIPprofile}(object)

\S4method{plotRegion}{ChIPprofile}(object = "ChIPprofile")
}
\description{
A function to plot regions
}
\author{
Thomas Carroll
}

