% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{makebedtable}
\alias{makebedtable}
\title{Make HTML pages for interval files or GRanges.}
\usage{
makebedtable(grangesObject, name, basedirectory)
}
\arguments{
\item{grangesObject}{A GRanges object.}

\item{name}{Character of the name for Interval HTML report.}

\item{basedirectory}{Character of the directory to which HTML report is writen.}
}
\value{
A character of file location for the Tracktables HTML Report
}
\description{
Creates HTML pages for interval files or GRanges (Tracktables Interval Report).
}
\examples{
data(Intervals)
htmlpage <- makebedtable(Intervals,"EBF_PeaksTable.html",getwd())
}
\author{
Thomas Carroll
}

