% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{makebedtable}
\alias{makebedtable}
\title{Make HTML page for interval files or GRanges}
\usage{
makebedtable(grangesObject, name, basedirectory)
}
\arguments{
\item{grangesObject}{A GRanges or object which may be parsed by ChIPQC GetGRanges method.}

\item{name}{Name for IGV session xml}

\item{basedirectory}{Directory to write IGV xml file}
}
\description{
Creates HTML page for interval files or GRanges
}
\examples{
\dontrun{
library(tracktables)

fileLocations <- system.file("extdata",package="tracktables")

bigwigs <- dir(fileLocations,pattern="*.bw",full.names=T)

intervals <- dir(fileLocations,pattern="*.bed",full.names=T)
intervalGRanges <- GetGRanges(intervals[1])
htmlpage <- makebedtable(intervalGRanges,"EBF_PeaksTable.html",getwd())
}
}
\author{
Thomas Carroll
}

