% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{MakeIGVSampleMetadata}
\alias{MakeIGVSampleMetadata}
\title{Make sample metadata file for use with IGV.}
\usage{
MakeIGVSampleMetadata(sampleMetadata, SampleSheet, igvdirectory)
}
\arguments{
\item{sampleMetadata}{A data.frame object with metadata information for samples.
First column must contain unique sample ids.}

\item{SampleSheet}{A data.frame of file locations. First column must contain the unique sample ids.}

\item{igvdirectory}{where igv xml files are located}
}
\description{
Creates sample metadata file for IGV from data.frames of user supplied metadata and
file locations
}
\examples{
\dontrun{
library(tracktables)

fileLocations <- system.file("extdata",package="tracktables")

bigwigs <- dir(fileLocations,pattern="*.bw",full.names=T)

intervals <- dir(fileLocations,pattern="*.bed",full.names=T)

bigWigMat <- cbind(gsub("_Example.bw","",basename(bigwigs)),
                   bigwigs)

intervalsMat <- cbind(gsub("_Peaks.bed","",basename(intervals)),
                      intervals)

fileSheet <- merge(bigWigMat,intervalsMat,all=T)

fileSheet <- as.matrix(cbind(fileSheet,NA))

colnames(fileSheet) <- c("SampleName","bigwig","interval","bam")

SampleSheet <- cbind(as.vector(fileSheet[,"SampleName"]),
                     c("EBF","H3K4me3","H3K9ac","RNAPol2"),
                     c("ProB","ProB","ProB","ProB"))

colnames(SampleSheet) <- c("SampleName","Antibody","Species")
MakeIGVSampleMetadata(SampleSheet,fileSheet,igvdirectory=getwd())
}
}
\author{
Thomas Carroll
}

