% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{MakeIGVSessionXML}
\alias{MakeIGVSessionXML}
\title{Make IGV session xml}
\usage{
MakeIGVSessionXML(SampleSheet, igvdirectory, XMLname, genomeName,
  locusName = "All")
}
\arguments{
\item{SampleSheet}{A samplesheet containing file locations}

\item{igvdirectory}{Directory for IGV}

\item{XMLname}{Name for IGV session xml}

\item{genomeName}{genome for IGV}

\item{locusName}{locus to display in igv on loading}
}
\description{
Creates session XML for IGV from Samplesheet and filesheet
}
\examples{
\dontrun{
library(tracktables)

fileLocations <- system.file("extdata",package="tracktables")

bigwigs <- dir(fileLocations,pattern="*.bw",full.names=T)

intervals <- dir(fileLocations,pattern="*.bed",full.names=T)

bigWigMat <- cbind(gsub("_Example.bw","",basename(bigwigs)),
                   bigwigs)

intervalsMat <- cbind(gsub("_Peaks.bed","",basename(intervals)),
                      intervals)

fileSheet <- merge(bigWigMat,intervalsMat,all=T)

fileSheet <- as.matrix(cbind(fileSheet,NA))

colnames(fileSheet) <- c("SampleName","bigwig","interval","bam")

SampleSheet <- cbind(as.vector(fileSheet[,"SampleName"]),
                     c("EBF","H3K4me3","H3K9ac","RNAPol2"),
                     c("ProB","ProB","ProB","ProB"))

colnames(SampleSheet) <- c("SampleName","Antibody","Species")
MakeIGVSessionXML(fileSheet,igvdirectory=getwd(),"Example","mm9")
}
}
\author{
Thomas Carroll
}

