% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{maketracktable}
\alias{maketracktable}
\title{Make HTML page for IGV sessions}
\usage{
maketracktable(fileSheet, SampleSheet, filename, basedirectory, genome)
}
\arguments{
\item{fileSheet}{A data frame or matrix containing sample file locations (e.g. BigWig locations).}

\item{SampleSheet}{A data frame or matrix containing sample metadata}

\item{filename}{Character string for tracktables HTML report}

\item{basedirectory}{Character string for directory for tracktables HTML report, IGV sessions and any interval files}

\item{genome}{Character string of genome for IGV}
}
\description{
Creates HTML page for IGV sessions
}
\examples{
fileLocations <- system.file("extdata",package="tracktables")

bigwigs <- dir(fileLocations,pattern="*.bw",full.names=TRUE)

intervals <- dir(fileLocations,pattern="*.bed",full.names=TRUE)

bigWigMat <- cbind(gsub("_Example.bw","",basename(bigwigs)),
                   bigwigs)

intervalsMat <- cbind(gsub("_Peaks.bed","",basename(intervals)),
                      intervals)

fileSheet <- merge(bigWigMat,intervalsMat,all=TRUE)

fileSheet <- as.matrix(cbind(fileSheet,NA))

colnames(fileSheet) <- c("SampleName","bigwig","interval","bam")

SampleSheet <- cbind(as.vector(fileSheet[,"SampleName"]),
                     c("EBF","H3K4me3","H3K9ac","RNAPol2"),
                     c("ProB","ProB","ProB","ProB"))

colnames(SampleSheet) <- c("SampleName","Antibody","Species")
  HTMLreport <- maketracktable(fileSheet,SampleSheet,
                               "IGV_Example",
                               basedirectory=getwd(),
                               "mm9")
}
\author{
Thomas Carroll
}

